Add new Visual Floss images and expand game data
Added scr_6.jpg to scr_10.jpg for Visual Floss. Updated AppAssets and AppConfig to support new games and descriptions for Visual Floss, including 'The Sacred Memory', 'A Day in a Life of a Dayfly', 'Incelthread', 'R/A\W', and 'NA-Re The Great Chain of Being'. Improved UI layout in main_page.dart, page_content_widget.dart, and showcase_widget.dart for better alignment and appearance.